+++ src/modules/Elsa.Workflows.Api/Endpoints/WorkflowDefinitions/Graph/Endpoint.cs
@@ -1,4 +1,5 @@
 using Elsa.Abstractions;
+using Elsa.Expressions.Contracts;
 using Elsa.Extensions;
 using Elsa.Workflows.Contracts;
 using Elsa.Workflows.Management;
@@ -9,7 +10,7 @@ using Microsoft.AspNetCore.Http;
 namespace Elsa.Workflows.Api.Endpoints.WorkflowDefinitions.Graph;
 
 [PublicAPI]
-internal class Graph(IWorkflowDefinitionService workflowDefinitionService, IApiSerializer apiSerializer) : ElsaEndpoint<Request>
+internal class Graph(IWorkflowDefinitionService workflowDefinitionService, IApiSerializer apiSerializer, IActivityRegistry activityRegistry, IExpressionDescriptorRegistry expressionDescriptorRegistry) : ElsaEndpoint<Request>
 {
     public override void Configure()
     {
@@ -29,7 +30,7 @@ internal class Graph(IWorkflowDefinitionService workflowDefinitionService, IApiS
 
         var parentNode = request.ParentNodeId == null ? workflowGraph.Root : workflowGraph.NodeIdLookup.TryGetValue(request.ParentNodeId, out var node) ? node : workflowGraph.Root;
         var serializerOptions = apiSerializer.GetOptions().Clone();
-        serializerOptions.Converters.Add(new ActivityNodeConverter(depth: 2));
+        serializerOptions.Converters.Add(new ActivityNodeConverter(activityRegistry, expressionDescriptorRegistry, depth: 2));
         await HttpContext.Response.WriteAsJsonAsync(parentNode, serializerOptions, cancellationToken);
     }
 }
