+++ src/modules/Elsa.Workflows.Core/Serialization/Converters/JsonIgnoreCompositeRootConverter.cs
@@ -2,15 +2,17 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Elsa.Expressions.Contracts;
 using Elsa.Workflows.Attributes;
 using Elsa.Workflows.Contracts;
+using Elsa.Workflows.Serialization.Helpers;
 
 namespace Elsa.Workflows.Serialization.Converters;
 
 /// <summary>
 /// Ignores properties with the <see cref="JsonIgnoreCompositeRootAttribute"/> attribute.
 /// </summary>
-public class JsonIgnoreCompositeRootConverter : JsonConverter<IActivity>
+public class JsonIgnoreCompositeRootConverter(IActivityRegistry activityRegistry, IExpressionDescriptorRegistry expressionDescriptorRegistry) : JsonConverter<IActivity>
 {
     /// <inheritdoc />
     public override IActivity Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
@@ -24,27 +26,34 @@ public class JsonIgnoreCompositeRootConverter : JsonConverter<IActivity>
     {
         writer.WriteStartObject();
 
-        var properties = value?.GetType().GetProperties(BindingFlags.Public | BindingFlags.Instance) ?? Array.Empty<PropertyInfo>();
-        
-        foreach (var property in properties)
+        if (value != null)
         {
-            if (property.GetCustomAttribute<JsonIgnoreAttribute>() != null)
-                continue;
-            
-            if (property.GetCustomAttribute<JsonIgnoreCompositeRootAttribute>() != null)
-                continue;
+            var properties = value.GetType().GetProperties(BindingFlags.Public | BindingFlags.Instance);
 
-            var propName = options.PropertyNamingPolicy?.ConvertName(property.Name) ?? property.Name;
-            writer.WritePropertyName(propName);
-            var input = property.GetValue(value);
-            
-            if (input == null)
+            foreach (var property in properties)
             {
-                writer.WriteNullValue();
-                continue;
+                if (property.GetCustomAttribute<JsonIgnoreAttribute>() != null)
+                    continue;
+
+                if (property.GetCustomAttribute<JsonIgnoreCompositeRootAttribute>() != null)
+                    continue;
+
+                var propName = options.PropertyNamingPolicy?.ConvertName(property.Name) ?? property.Name;
+                writer.WritePropertyName(propName);
+                var input = property.GetValue(value);
+
+                if (input == null)
+                {
+                    writer.WriteNullValue();
+                    continue;
+                }
+
+                JsonSerializer.Serialize(writer, input, options);
             }
-            
-            JsonSerializer.Serialize(writer, input, options);
+
+            var activityDescriptor = activityRegistry.Find(value.Type, value.Version);
+            if (activityDescriptor != null)
+                SyntheticPropertiesWriter.WriteSyntheticProperties(writer, value, activityDescriptor, expressionDescriptorRegistry, options);
         }
 
         writer.WriteEndObject();
