+++ src/modules/Elsa.Workflows.Core/Serialization/Converters/JsonIgnoreCompositeRootConverterFactory.cs
@@ -1,5 +1,6 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using Elsa.Expressions.Contracts;
 using Elsa.Workflows.Contracts;
 
 namespace Elsa.Workflows.Serialization.Converters;
@@ -7,7 +8,7 @@ namespace Elsa.Workflows.Serialization.Converters;
 /// <summary>
 /// A <see cref="JsonConverterFactory"/> that creates <see cref="JsonIgnoreCompositeRootConverter"/> instances.
 /// </summary>
-public class JsonIgnoreCompositeRootConverterFactory : JsonConverterFactory
+public class JsonIgnoreCompositeRootConverterFactory(IActivityRegistry activityRegistry, IExpressionDescriptorRegistry expressionDescriptorRegistry) : JsonConverterFactory
 {
     /// <inheritdoc />
     public override bool CanConvert(Type typeToConvert) => typeof(IActivity).IsAssignableFrom(typeToConvert);
@@ -15,6 +16,6 @@ public class JsonIgnoreCompositeRootConverterFactory : JsonConverterFactory
     /// <inheritdoc />
     public override JsonConverter CreateConverter(Type typeToConvert, JsonSerializerOptions options)
     {
-        return new JsonIgnoreCompositeRootConverter();
+        return new JsonIgnoreCompositeRootConverter(activityRegistry, expressionDescriptorRegistry);
     }
 }
